\name{N}
\alias{N}
\docType{data}
\title{ Effective population vector }
\description{
  A vector giving the effective population size for n populations. The effective population size is calculated as the total population divided by three.
}
\usage{data(N)}
\format{
  A 12 elements vector.
}
\details{
  This data comes for Swedlund et al. 1984.
}
\source{
  Swedlund, A. C., L. B. Jorde and J. H. Mielke. 1984. Population structure in the Connecticut valley. I. Marital migration. American Journal of Physical Anthropology 65: 61-70
}
\examples{
data(N)
}
\keyword{datasets}
