\name{P}
\alias{P}
\docType{data}
\title{ Column stochastic migration matrix }
\description{
  A column stochastic migration matrix for 12 populations.
}
\usage{data(P)}
\format{
  A 12 by 12 square matrix
}
\details{
  This data comes for Swedlund et al. 1984.
}
\source{
  Swedlund, A. C., L. B. Jorde and J. H. Mielke. 1984. Population structure in the Connecticut valley. I. Marital migration. American Journal of Physical Anthropology 65: 61-70
}
\examples{
data(P)
}
\keyword{datasets}
