\name{raw.mig}
\alias{raw.mig}
\docType{data}
\title{ A raw migration data }
\description{
  Made up raw dataset, created as if count data for marital migration were put into matrix form.
}
\usage{data(raw.mig)}
\format{
  A 4 by 4 square matrix.
}
\details{
  Completely made up for pedagogical purposes.
}

\examples{
data(raw.mig)
col.sto(raw.mig)
}
\keyword{datasets}
