\name{S}
\alias{S}
\docType{data}
\title{ Systematic pressure matrix }
\description{
  Systematic pressure matrix obtained by creating a square matix, where the non diagonal elements are all 0 and the diagonal elements are calculated as 1-Sk, where Sk is the systematic pressure for the k-th population.
}
\usage{data(S)}
\format{
  A 12 by 12 square matrix.
}
\details{
  This data comes for Swedlund et al. 1984.
}
\source{
  Swedlund, A. C., L. B. Jorde and J. H. Mielke. 1984. Population structure in the Connecticut valley. I. Marital migration. American Journal of Physical Anthropology 65: 61-70
}
\examples{
data(S)
}
\keyword{datasets}
